Refactor batch handler API.
Added laconiaContext when events are fired. It was
previously difficult to grab the Lambda context as
a local variable is needed to store the context.

Grouped batch parameter into one object that
consists of readerOptions and batchOptions.
This will be helpful when our API evolves i.e.
introduction of new option will not break our API.